backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tensorflow==2.16.1
Pillow==10.0.0
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.2